feat: rename files to workload
