+++ src/views/Workload/CreateView.vue
@@ -44,7 +44,7 @@
   import CreateFormBlock from '@/templates/create-form-block'
   import ContentBlock from '@/templates/content-block'
   import PageHeadingBlock from '@/templates/page-heading-block'
-  import FormFieldsCreateDomains from './FormFields/FormFieldsCreateDomains.vue'
+  import FormFieldsCreateDomains from './FormFields/FormFieldsCreateWorkload.vue'
   import ActionBarTemplate from '@/templates/action-bar-block/action-bar-with-teleport'
   import CopyDomainDialog from './Dialog/CopyDomainDialog.vue'
   import { useRoute, useRouter } from 'vue-router'
